runner/agent.rb: change log/pid file name completion.
The log_file and pid_file path were fixed values. Since the process name can
be changed based on the manifest.node_name attribute, it is reasonable that
those path names start with node_name as its default. The node_name is
determined after loading manifest file so that the location where daemonize()
called is changed.