change default prefetch option for rpc endpoint.
prefetch=1 used to be set for the messages which needs to be processed one by one. but
this makes more channels need per rpc endpoint.

i changed to use default channel, created at initialize_hook, for endpoints
that do not need to process each. it will be applied when new endpoint is created
without prefetch option.